Phase-2/backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "Todo backend service using FastAPI and SQLModel"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlmodel = "^0.0.16"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
uvicorn = "^0.24.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
flake8 = "^6.1.0"
mypy = "^1.7.1"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
skip-string-normalization = true

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",  # whitespace before ':'
    "W503",  # line break before binary operator
]